+++ CHANGELOG.md
@@ -55,6 +55,8 @@ changes.
 
 - `hydra-cardano-api` now published on [Cardano Haskell Packages (CHaP)](https://input-output-hk.github.io/cardano-haskell-packages/package/hydra-cardano-api-0.8.0/). [#504](https://github.com/input-output-hk/hydra/issues/504)
 
+- **BREAKING** `hydra-plutus` scripts changed.
+
 ## [0.8.0] - 2022-10-27
 
 - **BREAKING** Hydra keys now use the text envelope format.